Added a new module for GUI testing.
This module implements GUI class, which can be imported into tests and used as context manager.
This class enables interaction with GUI of the system under test.
Most importantly, the GUI class is responsible for taking screenshots and controlling the input devices like mouse and keybord.
The class implements a high level API with functions like click_on,
which takes the screenshot, finds a string in it and clicks on corresponding coordinate on the screen.